+++ packages/components/src/styles/link-button-styles.ts
@@ -12,13 +12,13 @@ import {
   hostHiddenStyles,
   hoverMediaQuery,
   prefersColorSchemeDarkMediaQuery,
+  SCALING_BASE_VALUE,
 } from './';
 import {
   borderRadiusSmall,
   borderWidthBase,
   fontLineHeight,
   frostedGlassStyle,
-  spacingStaticSmall,
   textSmallStyle,
 } from '@porsche-design-system/styles';
 
@@ -75,6 +75,7 @@ export const getLinkButtonStyles = (
   isDisabledOrLoading: boolean,
   hasSlottedAnchor: boolean,
   compact: boolean,
+  cssVariableInternalScaling: string,
   theme: Theme
 ): Styles => {
   const isPrimary = variant === 'primary';
@@ -93,8 +94,14 @@ export const getLinkButtonStyles = (
   const { focusColor } = getThemedColors(theme);
   const hasIcon = hasVisibleIcon(icon, iconSource) || hideLabel;
 
-  const paddingBlock = compact ? '4px' : '13px';
-  const paddingInline = compact ? '12px' : '26px';
+  const scalingVar = `var(${cssVariableInternalScaling}, ${compact ? 'calc(4 / 13)' : 1})`; // Compact mode needs to have 4px paddingBlock thus this scaling factor
+
+  const borderCompensation = variant === 'ghost' ? `+ ${borderWidthBase}` : ''; // Compensate for missing border in ghost variant (Fixes border backdrop-filter blur rendering issue in safari)
+
+  const paddingBlock = `calc(${scalingVar} * 0.8125 * ${SCALING_BASE_VALUE} ${borderCompensation})`; // 0.8125 * SCALING_BASE_VALUE corresponds to 13px
+  const paddingInline = `max(calc(${scalingVar} * 1.625 * ${SCALING_BASE_VALUE} ${borderCompensation}), ${variant === 'ghost' ? '6px' : '4px'})`; // 1.625 * SCALING_BASE_VALUE corresponds to 26px
+  const gap = `clamp(2px, calc(${scalingVar} * 0.5 * ${SCALING_BASE_VALUE}), 16px)`; // 0.5 * SCALING_BASE_VALUE corresponds to 8px
+  const iconMarginInlineStart = `clamp(-16px, calc(${scalingVar} * -0.5 * ${SCALING_BASE_VALUE}), -2px)`; // -0.5 * SCALING_BASE_VALUE corresponds to -8px
 
   return {
     '@global': {
@@ -103,6 +110,7 @@ export const getLinkButtonStyles = (
         ...addImportantToEachRule({
           verticalAlign: 'top',
           outline: 0, // custom element is able to delegate the focus
+          borderRadius: borderRadiusSmall,
           ...colorSchemeStyles,
           ...hostHiddenStyles,
         }),
@@ -112,23 +120,25 @@ export const getLinkButtonStyles = (
       display: 'flex',
       alignItems: 'flex-start',
       justifyContent: 'center',
-      width: '100%',
+      width: '100%', // Allows for setting a width on the host
+      minWidth: 'min-content', // Do not shrink beyond icon size + padding + border + label
       boxSizing: 'border-box',
       textAlign: 'start',
       WebkitAppearance: 'none', // iOS safari
       appearance: 'none',
       textDecoration: 'none',
-      border: `${borderWidthBase} solid ${borderColor}`,
+      ...textSmallStyle,
+      ...(variant === 'ghost'
+        ? { ...frostedGlassStyle, border: 'none' } // We can't use a border in the ghost variant due to rendering issues with backdrop-filter in safari
+        : { border: `${borderWidthBase} solid ${borderColor}` }),
       borderRadius: borderRadiusSmall,
       transform: 'translate3d(0,0,0)', // creates new stacking context (for slotted anchor + focus)
       backgroundColor,
-      ...(variant === 'ghost' && { ...frostedGlassStyle, backgroundClip: 'padding-box' }), // background color overlays border-color otherwise
       color: textColor,
-      ...textSmallStyle,
       transition: `${getTransition('background-color')}, ${getTransition('border-color')}, ${getTransition('color')}`,
       ...buildResponsiveStyles(hideLabel, (hideLabelValue: boolean) => ({
         padding: hideLabelValue ? paddingBlock : `${paddingBlock} ${paddingInline}`,
-        gap: hideLabelValue ? 0 : compact ? '6px' : spacingStaticSmall,
+        gap: hideLabelValue ? 0 : gap,
       })),
       ...(!hasSlottedAnchor && getFocusJssStyle(theme)),
       ...(!isDisabledOrLoading &&
@@ -155,7 +165,7 @@ export const getLinkButtonStyles = (
         width: fontLineHeight, // ensure space is already reserved until icon component is loaded (ssr)
         height: fontLineHeight, // ensure space is already reserved until icon component is loaded (ssr)
         ...buildResponsiveStyles(hideLabel, (hideLabelValue: boolean) => ({
-          marginInlineStart: hideLabelValue ? 0 : compact ? '-6px' : '-8px', // compensate white space of svg icon and optimize visual alignment
+          marginInlineStart: hideLabelValue ? 0 : iconMarginInlineStart, // compensate white space of svg icon and optimize visual alignment
         })),
       },
     }),